type casting 1

+++ day5/1.c
@@ -4,7 +4,7 @@ int main(){
 int x,y,n;
 printf("\nENTER X|Y|N ? ");
 scanf("%d%d%d",&x,&y,&n);
-int total;
+float total;
 total=(float)(x+y+n) / 3;
 printf("\n total = %i \n",total);
 }
